--- app/src/main/java/in/hridayan/ashell/ByteUtils.java
@@ -1,34 +1,0 @@
-package in.hridayan.ashell.utils;
-
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-
-/**
- * Created by xudong on 2/22/14.
- */
-public class ByteUtils {
-
-    public static byte[] concat(byte[]... arrays) {
-        // Determine the length of the result array
-        int totalLength = 0;
-        for (int i = 0; i < arrays.length; i++) {
-            totalLength += arrays[i].length;
-        }
-
-        // create the result array
-        byte[] result = new byte[totalLength];
-
-        // copy the source arrays into the result array
-        int currentIndex = 0;
-        for (int i = 0; i < arrays.length; i++) {
-            System.arraycopy(arrays[i], 0, result, currentIndex, arrays[i].length);
-            currentIndex += arrays[i].length;
-        }
-
-        return result;
-    }
-
-    public static final byte[] intToByteArray(int value) {
-        return ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putInt(value).array();
-    }
-}
