--- app/src/main/java/in/hridayan/ashell/Const.java
@@ -1,9 +1,0 @@
-package in.hridayan.ashell.utils;
-
-/**
- * Created by xudong on 2/7/14.
- */
-public class Const {
-    public static final String TAG = "flashbot";
-    public static final double PUSH_PERCENT = 0.5;
-}
