--- app/src/main/java/in/hridayan/ashell/MessageOtg.java
@@ -1,15 +1,0 @@
-package in.hridayan.ashell.utils;
-
-/**
- * Created by xudong on 2/7/14.
- */
-public class MessageOtg {
-    public static final int DEVICE_NOT_FOUND = 0;
-    public static final int CONNECTING = 1;
-    public static final int DEVICE_FOUND = 2;
-    public static final int FLASHING = 3;
-    public static final int INSTALLING_PROGRESS = 4;
-    public static final int PUSH_PART = 5;
-    public static final int PM_INST_PART = 6;
-    public static final String USB_PERMISSION = "hridayan.usb.permission";
-}
